fix: rollback removeFiles
